+++ Magic/Framework/SpellBook.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.IO;
+using System.Linq;
 using Magic.Framework.Spells;
 using Microsoft.Xna.Framework;
 using SpaceShared;
@@ -71,8 +72,11 @@ namespace Magic.Framework
         {
             var data = this.GetUpdatedData();
 
-            data.SelectedPrepared = (data.SelectedPrepared + 1) % data.Prepared.Count;
-            data.Save();
+            if (data.Prepared.Any())
+            {
+                data.SelectedPrepared = (data.SelectedPrepared + 1) % data.Prepared.Count;
+                data.Save();
+            }
         }
 
         /// <summary>Forget a known spell.</summary>