+++ Magic/docs/release-notes.md
@@ -1,7 +1,9 @@
 ﻿﻿[← back to readme](README.md)
 
 # Release notes
-## Upcoming release
+## 0.5.0
+Released 11 July 2021 for Stardew Valley 1.5. Updated by Pathoschild.
+
 * Added `magic_summary` console command to show troubleshooting info.
 * Added `help` documentation for console commands.
 * Fixed upgrading spells in Magic 0.4.0.